Added ProductDAO service and updated the README.
