+++ .gitignore
@@ -1,4 +1,4 @@
 /.bsp/
-/.g8/
+.g8/
 /.idea/
 target/
